refactor(handler): move schedule DTO mapper to schedules_dto.go (#1022)
Pure code motion: --color-moved clean, no exported-API or behaviour change.
Drops the now-unused "context" import from schedules.go (its only user,
repoPathFor, moved out). No positional comments needed fixing.
